Move testFailoverGenericSubscriberNextGenerationLoop to another class
Also move some methods to ConfigTest and do some more cleanup
